// apps/server/src/modules/upload/interfaces/upload-processor.interface.d.ts
export interface StandardPatentDocument {
    title: string;
    abstract: string;
    claims: string;
    keywords: string[];
    fullText: string;
}
export interface DirectTextInput {
    title: string;
    abstract: string;
    claims: string;
    keywords?: string[];
}
export interface ProcessDocumentFileInput {
    filename: string;
    mimetype: string;
    buffer: Buffer;
    size: number;
}
export interface ParsedDocumentResult {
    title?: string | undefined;
    bodyText: string;
    metadata?: Record<string, any> | undefined;
}
export interface IDocumentParser {
    parse(buffer: Buffer): Promise<ParsedDocumentResult>;
}
export interface IDocumentProcessorService {
    processFile(input: ProcessDocumentFileInput): Promise<StandardPatentDocument>;
    processDirectText(input: DirectTextInput): Promise<StandardPatentDocument>;
}
//# sourceMappingURL=upload-processor.interface.d.ts.map